requirements.txt
EAS2Text
EASGen
pydub
discord_webhook==0.17.0
requests
numpy
pyotp
qrcode
mysql-connector-python
colorama
flask
flask_sqlalchemy
mysqlclient
